# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "BankInfo",
      "version": "1.0.0",
      "title": "CompanyProfiles",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "servers": [
      {
        "url": "/"
      }
    ],
    "tags": [
      {
        "name": "Companies",
        "description": ""
      }
    ],
    "paths": {
      "/api/company/": {
        "get": {
          "tags": [
            "Companies List"
          ],
          "parameters": [
            {
                "name": "CompanyName",
                "in": "query",
                "required": false,
                "description": "name of company (String)",
                "type": "string"
            },
            {
                "name": "CompanyNumber",
                "in": "query",
                "required": false,
                "description": "number of company (Int)",
                "type": "int"
            },
            {
                "name": "CompanyStatusId",
                "in": "query",
                "required": false,
                "description": "status id of company (Int)",
                "type": "int"
            },{
                "name": "CompanyCategoryId",
                "in": "query",
                "required": false,
                "description": "category id of company (Int)",
                "type": "int"
            },
            {
                "name": "DistrictId",
                "in": "query",
                "required": false,
                "description": "DistrictId (Int)",
                "type": "int"
            },
            
            {
                "name": "BusinessActivityId",
                "in": "query",
                "required": false,
                "description": "BusinessActivityId (Int)",
                "type": "int"
            },
            {
                "name": "IncorporationDate",
                "in": "query",
                "required": false,
                "description": "IncorporationDate (Date YYYY-MM-DD) ",
                "type": "string"
            },
            {
                "name": "CountryOfOriginId",
                "in": "query",
                "required": false,
                "description": "CountryOfOriginId (Int)",
                "type": "int"
            },
            {
                "name": "CompanySizeId",
                "in": "query",
                "required": false,
                "description": "CompanySizeId (Int)",
                "type": "int"
            },
            {
                "name": "Exporter",
                "in": "query",
                "required": false,
                "description": "Exporter (Float)",
                "type": "int"
            },
            {
                "name": "HasPatent",
                "in": "query",
                "required": false,
                "description": "HasPatent (True/ False)",
                "type": "int"
            },
            {
                "name": "HasGrant",
                "in": "query",
                "required": false,
                "description": "HasGrant (True/ False)",
                "type": "bool"
            },
            {
                "name": "limit",
                "in": "query",
                "required": false,
                "description": "data per page (int)",
                "type": "int"
            },
            {
                "name": "offset",
                "in": "query",
                "required": false,
                "description": "page offset (int)",
                "type": "int"
            }
          ],
          "summary": "Returns All the companies" ,   
          "responses": {
            "200": {
              "description": "OK"
            },
            "400": {
              "description": "Failed. Misunderstood Request."
            },
            "404": {
              "description": "Failed. Book request not found."
            }
          }
      }
    },
      "/api/company/{comp_id}": {
        "parameters": [
          {
            "name": "comp_id",
            "in": "path",
            "required": true,
            "description": "ID of the company that we want to match",
            "type": "int"
          }
        ],
        "get": {
          "tags": [
            "Get single compnay"
          ],
          "summary": "Get compnay with given ID",
          "responses": {
            "200": {
              "description": "OK"
            },
            "400": {
              "description": "Failed. Misunderstood Request."
            },
            "404": {
              "description": "Failed. Book request not found."
            }
          }
        }
        }
      },
    "components": {
        "schemas": {
          "id": {
            "properties": {
              "uuid": {
                "type": "string"
              }
            }
          }
        }
      }
    }